+++ bridge/opentracing/mix_test.go
@@ -21,8 +21,8 @@ import (
 
 	ot "github.com/opentracing/opentracing-go"
 
+	"go.opentelemetry.io/otel"
 	otelglobal "go.opentelemetry.io/otel/api/global"
-	oteltrace "go.opentelemetry.io/otel/api/trace"
 	otelbaggage "go.opentelemetry.io/otel/internal/baggage"
 	"go.opentelemetry.io/otel/label"
 
@@ -142,8 +142,8 @@ func TestMixedAPIs(t *testing.T) {
 // simple test
 
 type simpleTest struct {
-	traceID oteltrace.ID
-	spanIDs []oteltrace.SpanID
+	traceID otel.TraceID
+	spanIDs []otel.SpanID
 }
 
 func newSimpleTest() *simpleTest {
@@ -177,11 +177,11 @@ func (st *simpleTest) noop(t *testing.T, ctx context.Context) context.Context {
 // current/active span test
 
 type currentActiveSpanTest struct {
-	traceID oteltrace.ID
-	spanIDs []oteltrace.SpanID
+	traceID otel.TraceID
+	spanIDs []otel.SpanID
 
-	recordedCurrentOtelSpanIDs []oteltrace.SpanID
-	recordedActiveOTSpanIDs    []oteltrace.SpanID
+	recordedCurrentOtelSpanIDs []otel.SpanID
+	recordedActiveOTSpanIDs    []otel.SpanID
 }
 
 func newCurrentActiveSpanTest() *currentActiveSpanTest {
@@ -229,10 +229,10 @@ func (cast *currentActiveSpanTest) runOTOtelOT(t *testing.T, ctx context.Context
 }
 
 func (cast *currentActiveSpanTest) recordSpans(t *testing.T, ctx context.Context) context.Context {
-	spanID := oteltrace.SpanFromContext(ctx).SpanContext().SpanID
+	spanID := otel.SpanFromContext(ctx).SpanContext().SpanID
 	cast.recordedCurrentOtelSpanIDs = append(cast.recordedCurrentOtelSpanIDs, spanID)
 
-	spanID = oteltrace.SpanID{}
+	spanID = otel.SpanID{}
 	if bridgeSpan, ok := ot.SpanFromContext(ctx).(*bridgeSpan); ok {
 		spanID = bridgeSpan.otelSpan.SpanContext().SpanID
 	}
@@ -423,7 +423,7 @@ func (bip *baggageItemsPreservationTest) addAndRecordBaggage(t *testing.T, ctx c
 
 type tracerMessTest struct {
 	recordedOTSpanTracers   []ot.Tracer
-	recordedOtelSpanTracers []oteltrace.Tracer
+	recordedOtelSpanTracers []otel.Tracer
 }
 
 func newTracerMessTest() *tracerMessTest {
@@ -475,7 +475,7 @@ func (tm *tracerMessTest) recordTracers(t *testing.T, ctx context.Context) conte
 		tm.recordedOTSpanTracers = append(tm.recordedOTSpanTracers, otSpan.Tracer())
 	}
 
-	otelSpan := oteltrace.SpanFromContext(ctx)
+	otelSpan := otel.SpanFromContext(ctx)
 	tm.recordedOtelSpanTracers = append(tm.recordedOtelSpanTracers, otelSpan.Tracer())
 	return ctx
 }
@@ -613,23 +613,23 @@ func generateBaggageKeys(key string) (otKey, otelKey string) {
 
 // helpers
 
-func checkTraceAndSpans(t *testing.T, tracer *internal.MockTracer, expectedTraceID oteltrace.ID, expectedSpanIDs []oteltrace.SpanID) {
+func checkTraceAndSpans(t *testing.T, tracer *internal.MockTracer, expectedTraceID otel.TraceID, expectedSpanIDs []otel.SpanID) {
 	expectedSpanCount := len(expectedSpanIDs)
 
 	// reverse spanIDs, since first span ID belongs to root, that
 	// finishes last
-	spanIDs := make([]oteltrace.SpanID, len(expectedSpanIDs))
+	spanIDs := make([]otel.SpanID, len(expectedSpanIDs))
 	copy(spanIDs, expectedSpanIDs)
 	reverse(len(spanIDs), func(i, j int) {
 		spanIDs[i], spanIDs[j] = spanIDs[j], spanIDs[i]
 	})
 	// the last finished span has no parent
-	parentSpanIDs := append(spanIDs[1:], oteltrace.SpanID{})
+	parentSpanIDs := append(spanIDs[1:], otel.SpanID{})
 
-	sks := map[oteltrace.SpanID]oteltrace.SpanKind{
-		{125}: oteltrace.SpanKindProducer,
-		{124}: oteltrace.SpanKindInternal,
-		{123}: oteltrace.SpanKindClient,
+	sks := map[otel.SpanID]otel.SpanKind{
+		{125}: otel.SpanKindProducer,
+		{124}: otel.SpanKindInternal,
+		{123}: otel.SpanKindClient,
 	}
 
 	if len(tracer.FinishedSpans) != expectedSpanCount {
@@ -660,12 +660,12 @@ func reverse(length int, swap func(i, j int)) {
 	}
 }
 
-func simpleTraceID() oteltrace.ID {
+func simpleTraceID() otel.TraceID {
 	return [16]byte{123, 42}
 }
 
-func simpleSpanIDs(count int) []oteltrace.SpanID {
-	base := []oteltrace.SpanID{
+func simpleSpanIDs(count int) []otel.SpanID {
+	base := []otel.SpanID{
 		{123},
 		{124},
 		{125},
@@ -685,7 +685,7 @@ func min(a, b int) int {
 
 func runOtelOTOtel(t *testing.T, ctx context.Context, name string, callback func(*testing.T, context.Context) context.Context) {
 	tr := otelglobal.Tracer("")
-	ctx, span := tr.Start(ctx, fmt.Sprintf("%s_Otel_OTOtel", name), oteltrace.WithSpanKind(oteltrace.SpanKindClient))
+	ctx, span := tr.Start(ctx, fmt.Sprintf("%s_Otel_OTOtel", name), otel.WithSpanKind(otel.SpanKindClient))
 	defer span.End()
 	ctx = callback(t, ctx)
 	func(ctx2 context.Context) {
@@ -693,7 +693,7 @@ func runOtelOTOtel(t *testing.T, ctx context.Context, name string, callback func
 		defer span.Finish()
 		ctx2 = callback(t, ctx2)
 		func(ctx3 context.Context) {
-			ctx3, span := tr.Start(ctx3, fmt.Sprintf("%sOtelOT_Otel_", name), oteltrace.WithSpanKind(oteltrace.SpanKindProducer))
+			ctx3, span := tr.Start(ctx3, fmt.Sprintf("%sOtelOT_Otel_", name), otel.WithSpanKind(otel.SpanKindProducer))
 			defer span.End()
 			_ = callback(t, ctx3)
 		}(ctx2)