+++ exporters/trace/jaeger/jaeger.go
@@ -22,8 +22,8 @@ import (
 
 	"google.golang.org/api/support/bundler"
 
+	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/api/global"
-	apitrace "go.opentelemetry.io/otel/api/trace"
 	"go.opentelemetry.io/otel/codes"
 	gen "go.opentelemetry.io/otel/exporters/trace/jaeger/internal/gen-go/jaeger"
 	"go.opentelemetry.io/otel/label"
@@ -92,8 +92,8 @@ func WithDisabled(disabled bool) Option {
 	}
 }
 
-// NewRawExporter returns a trace.Exporter implementation that exports
-// the collected spans to Jaeger.
+// NewRawExporter returns an OTel Exporter implementation that exports the
+// collected spans to Jaeger.
 //
 // It will IGNORE Disabled option.
 func NewRawExporter(endpointOption EndpointOption, opts ...Option) (*Exporter, error) {
@@ -151,14 +151,14 @@ func NewRawExporter(endpointOption EndpointOption, opts ...Option) (*Exporter, e
 
 // NewExportPipeline sets up a complete export pipeline
 // with the recommended setup for trace provider
-func NewExportPipeline(endpointOption EndpointOption, opts ...Option) (apitrace.TracerProvider, func(), error) {
+func NewExportPipeline(endpointOption EndpointOption, opts ...Option) (otel.TracerProvider, func(), error) {
 	o := options{}
 	opts = append(opts, WithDisabledFromEnv())
 	for _, opt := range opts {
 		opt(&o)
 	}
 	if o.Disabled {
-		return apitrace.NoopTracerProvider(), func() {}, nil
+		return otel.NewNoopTracerProvider(), func() {}, nil
 	}
 
 	exporter, err := NewRawExporter(endpointOption, opts...)
@@ -196,7 +196,8 @@ type Process struct {
 	Tags []label.KeyValue
 }
 
-// Exporter is an implementation of trace.SpanSyncer that uploads spans to Jaeger.
+// Exporter is an implementation of an OTel SpanSyncer that uploads spans to
+// Jaeger.
 type Exporter struct {
 	process  *gen.Process
 	bundler  *bundler.Bundler